mockgcp: refactory mock_http_roundtrip
This saves on LLM tokens!
